+++ tests/test_cobertura.py
@@ -25,7 +25,7 @@ def test_line_rate():
     assert cobertura.line_rate() == 0.9
 
 
-def test_line_rate_by_class():
+def test_line_rate_by_class_file():
     cobertura = make_cobertura()
     expected_line_rates = {
         'Main.java': 1.0,
@@ -34,9 +34,9 @@ def test_line_rate_by_class():
         'search/LinearSearch.java': 0.7142857142857143,
     }
 
-    for class_name in cobertura.classes():
-        assert cobertura.line_rate(class_name) == \
-            expected_line_rates[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.line_rate(class_filename) == \
+            expected_line_rates[class_filename]
 
 
 def test_branch_rate():
@@ -44,7 +44,7 @@ def test_branch_rate():
     assert cobertura.branch_rate() == 0.75
 
 
-def test_branch_rate_by_class():
+def test_branch_rate_by_class_file():
     cobertura = make_cobertura()
     expected_branch_rates = {
         'Main.java': 1.0,
@@ -53,12 +53,12 @@ def test_branch_rate_by_class():
         'search/LinearSearch.java': 0.6666666666666666,
     }
 
-    for class_name in cobertura.classes():
-        assert cobertura.branch_rate(class_name) == \
-            expected_branch_rates[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.branch_rate(class_filename) == \
+            expected_branch_rates[class_filename]
 
 
-def test_missed_statements_by_class_name():
+def test_missed_statements_by_class_file():
     cobertura = make_cobertura()
     expected_missed_statements = {
         'Main.java': [],
@@ -67,9 +67,9 @@ def test_missed_statements_by_class_name():
         'search/LinearSearch.java': [19, 24],
     }
 
-    for class_name in cobertura.classes():
-        assert cobertura.missed_statements(class_name) == \
-            expected_missed_statements[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.missed_statements(class_filename) == \
+            expected_missed_statements[class_filename]
 
 
 def test_list_packages():
@@ -82,7 +82,7 @@ def test_list_packages():
 def test_list_classes():
     cobertura = make_cobertura()
 
-    classes = cobertura.classes()
+    classes = cobertura.class_files()
     assert classes == [
         'Main.java',
         'search/BinarySearch.java',
@@ -101,8 +101,8 @@ def test_hit_lines__by_iterating_over_classes():
         'search/LinearSearch.java': [9, 13, 15, 16, 17],
     }
 
-    for class_name in cobertura.classes():
-        assert cobertura.hit_statements(class_name) == expected_lines[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.hit_statements(class_filename) == expected_lines[class_filename]
 
 
 def test_missed_lines():
@@ -115,8 +115,8 @@ def test_missed_lines():
         'search/LinearSearch.java': [19, 20, 21, 22, 23, 24],
     }
 
-    for class_name in cobertura.classes():
-        assert cobertura.missed_lines(class_name) == expected_lines[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.missed_lines(class_filename) == expected_lines[class_filename]
 
 
 def test_total_statements():
@@ -124,7 +124,7 @@ def test_total_statements():
     assert cobertura.total_statements() == 30
 
 
-def test_total_statements_by_class():
+def test_total_statements_by_class_file():
     cobertura = make_cobertura()
     expected_total_statements = {
         'Main.java': 11,
@@ -132,9 +132,9 @@ def test_total_statements_by_class():
         'search/ISortedArraySearch.java': 0,
         'search/LinearSearch.java': 7,
     }
-    for class_name in cobertura.classes():
-        assert cobertura.total_statements(class_name) == \
-            expected_total_statements[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.total_statements(class_filename) == \
+            expected_total_statements[class_filename]
 
 
 def test_total_misses():
@@ -142,7 +142,7 @@ def test_total_misses():
     assert cobertura.total_misses() == 3
 
 
-def test_total_misses_by_class():
+def test_total_misses_by_class_file():
     cobertura = make_cobertura()
     expected_total_misses = {
         'Main.java': 0,
@@ -150,9 +150,9 @@ def test_total_misses_by_class():
         'search/ISortedArraySearch.java': 0,
         'search/LinearSearch.java': 2,
     }
-    for class_name in cobertura.classes():
-        assert cobertura.total_misses(class_name) == \
-            expected_total_misses[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.total_misses(class_filename) == \
+            expected_total_misses[class_filename]
 
 
 def test_total_hits():
@@ -160,7 +160,7 @@ def test_total_hits():
     assert cobertura.total_hits() == 27
 
 
-def test_total_hits_by_class():
+def test_total_hits_by_class_file():
     cobertura = make_cobertura()
     expected_total_misses = {
         'Main.java': 11,
@@ -168,22 +168,9 @@ def test_total_hits_by_class():
         'search/ISortedArraySearch.java': 0,
         'search/LinearSearch.java': 5,
     }
-    for class_name in cobertura.classes():
-        assert cobertura.total_hits(class_name) == \
-            expected_total_misses[class_name]
-
-
-def test_filename():
-    cobertura = make_cobertura()
-    expected_filenames = {
-        'Main.java': 'Main.java',
-        'search/BinarySearch.java': 'search/BinarySearch.java',
-        'search/ISortedArraySearch.java': 'search/ISortedArraySearch.java',
-        'search/LinearSearch.java': 'search/LinearSearch.java',
-    }
-    for class_name in cobertura.classes():
-        assert cobertura.filename(class_name) == \
-            expected_filenames[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.total_hits(class_filename) == \
+            expected_total_misses[class_filename]
 
 
 def test_filepath():
@@ -195,12 +182,12 @@ def test_filepath():
         'search/ISortedArraySearch.java': 'foo/bar/baz/search/ISortedArraySearch.java',
         'search/LinearSearch.java': 'foo/bar/baz/search/LinearSearch.java',
     }
-    for class_name in cobertura.classes():
-        assert cobertura.filepath(class_name) == \
-            expected_filepaths[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.filepath(class_filename) == \
+            expected_filepaths[class_filename]
 
 
-def test_class_source__sources_not_found():
+def test_class_file_source__sources_not_found():
     from pycobertura.cobertura import Line
     cobertura = make_cobertura('tests/cobertura.xml')
     expected_sources = {
@@ -209,8 +196,8 @@ def test_class_source__sources_not_found():
         'search/ISortedArraySearch.java': [Line(0, 'tests/search/ISortedArraySearch.java not found', None, None)],
         'search/LinearSearch.java': [Line(0, 'tests/search/LinearSearch.java not found', None, None)],
     }
-    for class_name in cobertura.classes():
-        assert cobertura.class_source(class_name) == expected_sources[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.class_file_source(class_filename) == expected_sources[class_filename]
 
 
 def test_line_statuses():
@@ -236,12 +223,12 @@ def test_line_statuses():
             (6, False)
         ],
     }
-    for class_name in cobertura.classes():
-        assert cobertura.line_statuses(class_name) == \
-            expected_line_statuses[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.line_statuses(class_filename) == \
+            expected_line_statuses[class_filename]
 
 
-def test_class_source__sources_found():
+def test_class_file_source__sources_found():
     from pycobertura.cobertura import Line
     cobertura = make_cobertura('tests/dummy.source1/coverage.xml')
     expected_sources = {
@@ -267,6 +254,6 @@ def test_class_source__sources_found():
             Line(6, '    pass\n', False, None)
         ],
     }
-    for class_name in cobertura.classes():
-        assert cobertura.class_source(class_name) == \
-            expected_sources[class_name]
+    for class_filename in cobertura.class_files():
+        assert cobertura.class_file_source(class_filename) == \
+               expected_sources[class_filename]